Create README.md
I DO believe in miracles and am a constant believer in 'TIMING.' These are trying days for our species. Unless there is soon a positive and sensible turn-around in my mostly completed site, I shall have to 'Throw in the Towel,' as they say....Any response to the above better be for real! I have had it with posers!